+++ pipeline/docker/java-s3-copy/scripts/copy_to_s3.sh
@@ -2,7 +2,10 @@
 # copy_to_s3
 
 cd $1
-tar -cvf vis.tar vis_*
 
-java -classpath /chiles02/awsChiles02.jar org.icrar.awsChiles02.copyS3.CopyFileToS3 -aws_access_key_id $3 -aws_secret_access_key $4 $2 vis.tar
-rm vis.tar
+if [ -d vis_$5~$6 ]; then
+    tar -cvf vis.tar vis_$5~$6
+
+    java -classpath /chiles02/awsChiles02.jar org.icrar.awsChiles02.copyS3.CopyFileToS3 -aws_access_key_id $3 -aws_secret_access_key $4 $2 vis.tar
+    rm vis.tar
+fi
